Consolidation of MakeApp and TestApp files
Moved TestApp (the template) into  the MakeApp directory.  This allows the template to be installed along with MakeApp so that MakeApp can produce a new application from its protected data...   The only real requirement is that the Library directory be near (if not in the same directory) as the new project.